sc_data_interface/requirements.txt
attrs==21.4.0
cycler==0.11.0
dnspython==2.2.1
fonttools==4.31.2
importlib-resources==5.6.0
jsonschema==4.4.0
kafka-python==2.0.2
kiwisolver==1.4.2
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pymongo==4.1.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
six==1.16.0
zipp==3.8.0
